refactor pt.2
